Use caml_alloc_custom_mem instead of "manually" tracking the C heap
caml_alloc_custom_mem does exist specifically for custom block using out of heap
memory. The GC has an heuristic depending on the out of heap memory used by the
custom block. It is better to rely on the heuristic than the parameter used/max
to instrument the GC to perform a collection.